Umstrukturierung E-Mail-Funktionalität

+++ app/Http/Controllers/App/InboxController.php
@@ -12,18 +12,24 @@ use App\Models\Contact;
 use App\Models\DropboxInbox;
 use App\Models\Project;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Inertia\Response;
 use Throwable;
 
 class InboxController extends Controller
 {
-    public function index($mail = null): Response
+    public function index(?int $mail = null): Response
     {
+
+        if (!Auth::user()->is_admin) {
+            abort(403);
+        }
+
         if ($mail) {
             $mail = DropboxInbox::query()->with('dropbox')->where('id', $mail)->first();
             if ($mail) {
-                if (! $mail->seen_at) {
+                if (!$mail->seen_at) {
                     $mail->seen_at = now();
                     $mail->save();
                 }
@@ -47,18 +53,31 @@ class InboxController extends Controller
     /**
      * @throws Throwable
      */
-    public function import($mail): RedirectResponse
+    public function import(int $mail): RedirectResponse
     {
+        if (!Auth::user()->is_admin) {
+            abort(403);
+        }
+
         $mail = DropboxInbox::query()->with('dropbox')->where('id', $mail)->first();
+
+        if (!$mail) {
+            abort(404);
+        }
+
         DropboxImportJob::dispatch($mail);
 
-        return redirect()->route('app.inbox.index');
+        return redirect()->route('admin.inbox.index');
     }
 
     public function destroy(DropboxInbox $mail): RedirectResponse
     {
+        if (!Auth::user()->is_admin) {
+            abort(403);
+        }
+
         $mail->delete();
 
-        return redirect()->route('app.inbox.index');
+        return redirect()->route('admin.inbox.index');
     }
 }
